Resolved conflicts in 8 files:
- server/scripts/calculate-aggregate-scores.ts: Kept parameterized COALESCE approach
- server/src/lib/aggregate-score.ts: Kept improved BASE_WEIGHTS comment
- server/src/routes/movie.test.ts: Combined extra test case with cleaner types
- server/src/routes/shows.test.ts: Used cleaner mock comment
- src/components/common/AggregateScore.test.tsx: Kept floating-point comment, simpler undefined
- src/components/common/AggregateScore.tsx: Used broader undefined-accepting types
- src/pages/MoviePage.tsx: Kept simpler unconditional AggregateScore rendering
- src/pages/ShowPage.tsx: Kept simpler unconditional AggregateScore rendering

+++ server/src/routes/admin/data-quality.ts
@@ -142,7 +142,7 @@ router.get("/future-deaths", async (req: Request, res: Response): Promise<void> 
         tmdb_id,
         deathday,
         birthday,
-        popularity,
+        popularity::float,
         CASE
           WHEN deathday > CURRENT_DATE THEN 'future_date'
           WHEN birthday IS NOT NULL AND deathday < birthday THEN 'before_birth'
@@ -350,7 +350,7 @@ router.get("/uncertain-deaths", async (req: Request, res: Response): Promise<voi
         a.name,
         a.tmdb_id,
         a.deathday,
-        a.popularity,
+        a.popularity::float,
         adc.circumstances
       FROM actors a
       JOIN actor_death_circumstances adc ON a.id = adc.actor_id
